Update example notebook Rename the library mplc Clean repo tree
