+++ mplc/__init__.py
@@ -1,5 +1,6 @@
 from . import constants  # noqa
 from . import contributivity  # noqa
+from . import dataset  # noqa
 from . import datasets  # noqa
 from . import multi_partner_learning  # noqa
 from . import scenario  # noqa
