+++ mplc/datasets/dataset_cifar10.py
@@ -14,8 +14,7 @@ from keras.utils import to_categorical
 from loguru import logger
 from sklearn.model_selection import train_test_split
 
-from . import dataset
-from .. import constants
+from .. import constants, dataset
 
 # Init dataset-specific variables
 input_shape = (32, 32, 3)
