+++ setup.py
@@ -3,11 +3,11 @@ from setuptools import setup, find_packages
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
-with open("subtest/requirements.txt", "r") as fh:
+with open("mplc/requirements.txt", "r") as fh:
     requirements = fh.read()
 
 setup(
-    name="subtest",
+    name="mplc",
     author="SubstraFoundation",
     author_email="contact@substra.org",
     description="A distributed learning contributivity package",
@@ -29,3 +29,4 @@ setup(
     install_requires=requirements,
     python_requires='>=3.6',
 )
+
