--- shapley_value/TestCases/TC2.txt
@@ -1,2 +1,0 @@
-N = 5
-1, 4, 3, 1, 5, 3, 3, 2, 1, 3, 2, 5, 6, 4, 6, 6, 4, 2, 4, 0, 3, 0, 7, 5, 6, 8, 8, 4, 1, 1, 9
