--- shapley_value/TestCases/TC3.txt
@@ -1,2 +1,0 @@
-N = 7
-7, 6, 1, 3, 1, 7, 2, 3, 6, 1, 5, 4, 5, 7, 5, 4, 6, 0, 7, 1, 8, 8, 6, 6, 8, 8, 8, 4, 7, 5, 9, 2, 2, 8, 9, 7, 3, 6, 1, 2, 9, 3, 1, 9, 4, 7, 8, 4, 5, 0, 3, 6, 1, 0, 6, 3, 2, 0, 6, 1, 5, 5, 4, 3, 6, 9, 10, 2, 4, 1, 0, 10, 9, 4, 0, 10, 1, 2, 6, 9, 7, 1, 4, 9, 3, 9, 2, 10, 3, 7, 6, 8, 9, 8, 9, 4, 5, 9, 6, 11, 3, 10, 2, 10, 11, 9, 5, 2, 9, 0, 4, 8, 1, 8, 2, 1, 9, 5, 4, 10, 12, 8, 6, 4, 4, 1, 3
