--- shapley_value/TestCases/TC4.txt
@@ -1,2 +1,0 @@
-N = 9
-3, 6, 7, 5, 3, 5, 6, 2, 9, 1, 8, 7, 5, 3, 7, 4, 9, 10, 2, 0, 10, 8, 5, 0, 4, 6, 0, 10, 3, 10, 10, 7, 10, 3, 7, 9, 7, 8, 2, 10, 7, 10, 4, 1, 0, 1, 2, 8, 11, 6, 5, 10, 7, 8, 8, 9, 2, 2, 4, 1, 9, 9, 8, 3, 8, 5, 10, 9, 1, 5, 0, 3, 6, 9, 2, 3, 2, 8, 3, 2, 7, 0, 0, 2, 8, 0, 3, 11, 2, 7, 0, 0, 8, 0, 7, 4, 10, 6, 6, 11, 3, 10, 2, 10, 11, 9, 5, 2, 9, 0, 4, 8, 1, 8, 2, 1, 9, 5, 4, 11, 4, 9, 3, 1, 9, 11, 9, 11, 9, 6, 0, 5, 0, 5, 9, 12, 3, 5, 9, 12, 0, 5, 0, 3, 4, 4, 9, 5, 1, 5, 5, 1, 2, 0, 7, 8, 3, 7, 5, 5, 1, 5, 0, 3, 10, 9, 2, 3, 2, 0, 2, 4, 7, 2, 7, 12, 8, 5, 4, 10, 12, 11, 0, 1, 12, 9, 10, 2, 3, 2, 10, 4, 9, 10, 10, 8, 6, 12, 11, 10, 0, 2, 1, 7, 7, 10, 8, 2, 3, 2, 1, 4, 0, 3, 6, 1, 12, 5, 6, 3, 9, 3, 9, 5, 2, 6, 2, 10, 8, 3, 8, 10, 7, 11, 6, 3, 11, 2, 6, 3, 4, 9, 7, 6, 0, 2, 8, 1, 7, 1, 4, 3, 6, 3, 10, 7, 8, 10, 2, 0, 10, 11, 12, 12, 2, 2, 0, 9, 8, 2, 2, 0, 4, 1, 5, 10, 10, 1, 3, 9, 5, 8, 11, 4, 10, 9, 11, 5, 5, 12, 3, 13, 7, 2, 12, 9, 4, 10, 2, 10, 12, 4, 10, 2, 4, 13, 11, 12, 1, 0, 7, 4, 8, 2, 7, 5, 12, 2, 8, 1, 0, 11, 13, 6, 11, 9, 13, 13, 5, 2, 9, 3, 4, 3, 3, 6, 1, 0, 4, 0, 12, 10, 2, 7, 12, 9, 10, 8, 12, 4, 10, 10, 13, 9, 2, 9, 4, 2, 8, 7, 2, 2, 10, 4, 3, 11, 11, 4, 12, 13, 4, 8, 9, 5, 1, 8, 0, 9, 2, 12, 11, 4, 13, 10, 11, 7, 14, 13, 1, 14, 6, 9, 4, 7, 7, 11, 14, 1, 12, 3, 3, 4, 7, 3, 3, 3, 6, 3, 13, 8, 3, 13, 13, 8, 0, 9, 7, 6, 8, 9, 5, 6, 3, 2, 14, 10, 13, 13, 3, 2, 8, 7, 6, 0, 2, 1, 10, 8, 12, 9, 1, 0, 14, 6, 0, 14, 8, 7, 5, 8, 1, 2, 14, 12, 4, 13, 14, 2, 3, 3, 5, 11, 2, 3, 4, 15, 7, 6, 3, 12, 7, 15, 8, 13, 1, 4, 13, 13, 15, 15, 9, 4, 6, 3, 5, 1, 7, 4, 4, 3, 7, 8, 13, 6, 2, 12, 5, 9, 2, 8, 5, 7, 0, 13, 9, 7, 11, 7, 8, 9, 8
