--- shapley_value/TestCases/TC5.txt
@@ -1,2 +1,0 @@
-N = 10
-6, 10, 6, 2, 1, 4, 0, 6, 3, 1, 2, 7, 2, 7, 11, 10, 0, 6, 4, 4, 11, 8, 3, 9, 2, 2, 2, 7, 7, 7, 5, 6, 6, 6, 5, 3, 1, 4, 7, 10, 9, 9, 9, 11, 8, 1, 2, 8, 11, 6, 5, 10, 7, 8, 8, 3, 11, 6, 6, 10, 10, 2, 10, 11, 1, 4, 4, 11, 11, 0, 12, 2, 0, 2, 3, 5, 9, 10, 7, 1, 2, 10, 12, 2, 1, 3, 7, 0, 9, 0, 10, 7, 4, 9, 7, 5, 3, 12, 6, 1, 1, 7, 3, 2, 11, 9, 9, 8, 8, 5, 9, 12, 3, 10, 3, 6, 1, 10, 8, 10, 12, 8, 6, 4, 4, 1, 11, 9, 2, 6, 0, 5, 0, 5, 9, 12, 3, 5, 9, 12, 0, 5, 0, 3, 4, 4, 9, 5, 1, 5, 5, 1, 2, 0, 7, 8, 3, 7, 5, 5, 1, 5, 0, 3, 10, 9, 2, 3, 2, 0, 2, 4, 7, 2, 7, 7, 3, 1, 3, 0, 2, 11, 7, 6, 9, 9, 12, 1, 0, 9, 7, 11, 1, 9, 10, 12, 5, 3, 4, 1, 6, 6, 11, 9, 13, 2, 0, 2, 4, 2, 0, 4, 13, 5, 11, 6, 0, 7, 6, 1, 0, 13, 10, 1, 6, 6, 11, 9, 8, 0, 11, 12, 4, 6, 5, 1, 7, 6, 4, 9, 8, 2, 13, 5, 8, 8, 12, 6, 1, 4, 7, 1, 1, 2, 0, 7, 8, 10, 2, 0, 10, 11, 12, 12, 2, 2, 0, 9, 8, 2, 2, 0, 4, 1, 5, 10, 10, 1, 3, 9, 5, 8, 11, 4, 10, 9, 11, 5, 5, 12, 3, 13, 7, 2, 12, 9, 4, 10, 2, 10, 12, 4, 10, 2, 4, 13, 11, 12, 1, 0, 7, 4, 8, 2, 7, 5, 12, 2, 8, 1, 0, 11, 13, 6, 11, 9, 13, 13, 5, 2, 9, 3, 4, 3, 3, 6, 1, 0, 4, 0, 12, 10, 2, 7, 12, 9, 10, 8, 12, 4, 10, 10, 13, 9, 2, 9, 4, 2, 8, 7, 2, 2, 10, 4, 3, 11, 11, 4, 12, 13, 4, 8, 9, 5, 1, 8, 0, 9, 2, 12, 11, 10, 9, 11, 3, 7, 14, 13, 1, 14, 6, 9, 4, 7, 7, 11, 14, 1, 12, 3, 3, 4, 7, 3, 3, 3, 6, 3, 13, 8, 3, 13, 13, 8, 0, 9, 7, 6, 8, 9, 5, 6, 3, 2, 14, 10, 13, 13, 3, 2, 8, 7, 6, 0, 2, 1, 10, 8, 12, 9, 1, 0, 14, 6, 0, 14, 8, 7, 5, 8, 1, 2, 14, 12, 4, 13, 14, 2, 3, 3, 5, 11, 2, 3, 4, 11, 5, 14, 4, 2, 0, 12, 9, 14, 4, 9, 5, 4, 8, 10, 12, 10, 13, 11, 14, 9, 2, 13, 12, 5, 8, 9, 9, 10, 12, 5, 6, 9, 11, 2, 11, 12, 0, 12, 3, 11, 6, 9, 0, 0, 11, 12, 2, 9, 0, 1, 4, 2, 14, 8, 0, 8, 2, 1, 10, 6, 6, 9, 1, 2, 3, 4, 6, 10, 2, 10, 6, 0, 11, 13, 7, 7, 2, 1, 9, 10, 2, 5, 12, 9, 5, 4, 9, 14, 5, 11, 5, 3, 12, 13, 13, 1, 10, 4, 11, 12, 6, 10, 4, 2, 0, 4, 2, 3, 5, 3, 13, 8, 8, 2, 9, 13, 7, 3, 12, 4, 6, 9, 0, 4, 8, 13, 5, 3, 9, 8, 7, 1, 3, 3, 10, 4, 7, 4, 14, 13, 7, 4, 13, 7, 6, 7, 12, 5, 2, 1, 2, 0, 11, 2, 4, 11, 7, 1, 6, 1, 2, 5, 9, 5, 8, 5, 1, 1, 1, 7, 6, 11, 4, 12, 2, 4, 2, 13, 2, 6, 12, 3, 1, 13, 0, 0, 0, 11, 7, 15, 9, 0, 11, 7, 15, 14, 4, 8, 10, 1, 3, 1, 12, 7, 13, 14, 11, 15, 11, 14, 5, 8, 1, 6, 5, 2, 6, 6, 13, 14, 5, 6, 14, 1, 14, 13, 15, 2, 6, 9, 3, 9, 10, 15, 0, 7, 14, 12, 6, 9, 10, 11, 1, 11, 1, 7, 13, 8, 13, 11, 6, 2, 1, 4, 3, 15, 2, 2, 2, 8, 11, 5, 1, 5, 5, 1, 12, 3, 13, 2, 12, 7, 13, 14, 3, 15, 5, 0, 7, 2, 11, 13, 4, 13, 1, 8, 12, 3, 10, 14, 11, 6, 4, 12, 11, 9, 14, 8, 12, 11, 10, 8, 3, 8, 6, 6, 7, 11, 6, 14, 13, 2, 11, 2, 15, 12, 10, 11, 0, 4, 10, 11, 10, 14, 8, 6, 7, 6, 14, 3, 1, 8, 11, 4, 0, 2, 10, 7, 13, 1, 5, 11, 3, 0, 13, 2, 13, 7, 13, 13, 11, 7, 8, 6, 5, 0, 12, 12, 6, 10, 15, 8, 2, 11, 12, 3, 13, 7, 10, 10, 8, 0, 5, 11, 0, 2, 13, 13, 9, 10, 10, 4, 2, 3, 7, 7, 8, 4, 16, 7, 3, 16, 13, 15, 11, 12, 11, 8, 7, 8, 15, 12, 16, 3, 7, 11, 1, 11, 11, 12, 0, 16, 16, 2, 2, 6, 0, 2, 1, 7, 0, 4, 7, 4, 3, 9, 16, 5, 9, 7, 13, 7, 10, 3, 1, 9, 15, 10, 3, 0, 6, 11, 16, 5, 4, 10, 3, 5, 3, 12, 3, 3, 0, 1, 7, 11, 11, 6, 16, 11, 4, 4, 1, 15, 15, 10, 7, 4, 4, 1, 4, 10, 13, 12, 6, 0, 5, 0, 13, 8, 13, 8, 2, 4, 9, 0, 15, 11, 6, 5, 5, 2, 0, 14, 8, 16, 8, 6, 3, 12, 7, 17, 3, 4, 10, 7, 17, 1, 6, 14, 14, 9, 12, 11, 16, 4, 2, 4, 14, 11, 16, 9, 1, 4, 6, 7, 14, 11, 17, 7, 4, 1, 6, 5, 5, 16, 13, 4, 15, 17, 17, 9, 8, 11, 2, 4, 2, 7, 6, 10, 10, 14, 6, 12, 11, 4, 15
