--- shapley_value/TestCases/divide_dollar
@@ -1,2 +1,0 @@
-N = 3
-0, 0, 0, 30, 0, 0, 30
