--- shapley_value/TestCases/samplegame1
@@ -1,2 +1,0 @@
-N = 4
-1,1,2,1,3,4,4,3,4,4,6,7,6,6,8
