--- shapley_value/TestCases/samplegame2
@@ -1,2 +1,0 @@
-N = 3
-1,1,1,3,4,4,7
