--- shapley_value/TestCases/samplegame3
@@ -1,2 +1,0 @@
-N = 3
-0,0,0,30,30,30,30
